+++ lib/src/parent_progress.dart
@@ -1,10 +1,25 @@
 import 'package:flutter/foundation.dart';
 import 'package:parent_progress/src/child_progress.dart';
-
 import 'debug_helpers.dart';
 
-/// A class that aggregates multiple progress progresses and calculates a weighted total progress.
-/// This class uses `_flexFactors` to determine the weight of each child progress's contribution to the total progress.
+/// A class that aggregates multiple child progress instances and calculates a total progress
+/// with weighted contributions from each child. This class is particularly useful in scenarios
+/// where multiple concurrent operations contribute to a single overarching goal, such as
+/// downloading multiple files, processing multiple data streams, or completing stages of a
+/// complex task. Inherits functionalities from `ChildProgress`.
+///
+/// Features include:
+/// - Dynamic weighting of each child's progress contribution, allowing for flexible prioritization
+///   of tasks based on their importance or impact on the overall process.
+/// - Real-time calculation and notification of total progress, providing a comprehensive view
+///   of the combined progress of all children.
+/// - Easy integration with UI components to display collective progress, enhancing user
+///   experience by showing unified progress details.
+///
+/// The class uses `_flexFactors` to determine the weight of each child progress's contribution
+/// to the total progress, offering a customizable approach to progress aggregation. This mechanism
+/// ensures that some tasks can be prioritized higher than others, depending on the specified
+/// weights, making the class adaptable to a wide range of applications.
 class ParentProgress extends ChildProgress {
   int _percentage = 0;
   final List<int>
@@ -17,11 +32,10 @@ class ParentProgress extends ChildProgress {
   final List<VoidCallback> _listeners =
       []; // List to store listeners for cleanup.
 
-  /// Initializes a new ParentProgress with a list of child percentage notifiers and their corresponding flex factors.
-  /// Throws ArgumentError if the lists' lengths do not match.
-  ///
-  /// The `percentageNotifiers` list should include ValueNotifier<int> from each child progress,
+  /// Initializes a new ParentProgress with a list of child progress objects and their corresponding flex factors.
+  /// Each child progress's `percentageNotifier` is included implicitly via the `children` list,
   /// and `flexFactors` determines the importance or weight of each child's progress.
+  /// Throws an ArgumentError if the lists' lengths do not match.
   ParentProgress(
     List<ChildProgress> children,
     List<int> flexFactors, {
@@ -37,9 +51,16 @@ class ParentProgress extends ChildProgress {
     _initializeListeners();
   }
 
+  /// Returns a list of flex factors that determine the weight of each child progress's contribution.
   List<int> get getFlexFactors => _flexFactors;
+
+  /// Returns a list of child progress instances, each representing a progress tracking component.
   List<ChildProgress> get getChildren => _children;
+
+  /// Returns a list of `ValueNotifier<int>` from each child progress, used to notify listeners of percentage changes.
   List<ValueNotifier<int>> get getSlices => _slices;
+
+  /// Returns a list of progress weights calculated from each child's current percentage, influenced by the associated flex factor.
   List<double> get getProgressWeights => _progressWeights;
 
   /// Sets up listeners on each child progress's ValueNotifier to update the parent progress.
@@ -53,8 +74,9 @@ class ParentProgress extends ChildProgress {
         '${uniqueName != null ? '${uniqueName!}: ' : ''}ParentProgress initialized.');
   }
 
-  /// Updates the progress contribution from a specific child progress based on its new percentage.
-  /// Validates the index and percentage range, and recalculates the total progress.
+  /// Updates the progress contribution from a specific child based on its new percentage.
+  /// Validates the provided index against the range of child progress instances and the percentage value must be between 0 and 100.
+  /// Throws an `IndexError` if the index is out of bounds or an `ArgumentError` if the percentage is out of valid range.
   void _updateSliceProgress(int sliceIndex, int newPercentage) {
     if (sliceIndex < 0 || sliceIndex >= _flexFactors.length) {
       throw IndexError.withLength(sliceIndex, _flexFactors.length);
@@ -78,7 +100,8 @@ class ParentProgress extends ChildProgress {
         "${uniqueName != null ? '${uniqueName!}: ' : ''}Parent percentage updated: $_percentage");
   }
 
-  /// Cleans up by removing listeners and disposing of the notifier.
+  /// Cleans up by removing all listeners from child progress notifiers and disposing of the total percentage notifier.
+  /// It's essential to call this method to free up resources when the `ParentProgress` instance is no longer needed.
   void dispose() {
     for (int i = 0; i < _slices.length; i++) {
       _slices[i].removeListener(_listeners[i]);